refactor(lib/oauth): X provider as UniversalOAuthProvider subclass (base PKCE)
